Almost done except for a few todos
